+++ quick_sort.h
@@ -1,8 +1,8 @@
 #ifndef __QUICK_SORT__
 #define __QUICK_SORT__
 
-void swap(float *a, float *b);
-int partition(float array[], int low, int high);
-void quickSort(float array[], int low, int high);
+void swap(int *a, int *b);
+int partition(int array[], int low, int high);
+void quickSort(int array[], int low, int high);
 
 #endif
